Moved `ViewportBlueprint`, `ContainerBlueprint`, and friends
