+++ crates/re_space_view_text_log/Cargo.toml
@@ -33,4 +33,3 @@ re_viewer_context.workspace = true
 
 egui_extras.workspace = true
 egui.workspace = true
-itertools.workspace = true
