+++ crates/re_viewer/Cargo.toml
@@ -92,7 +92,6 @@ eframe = { workspace = true, default-features = false, features = [
   "puffin",
   "wgpu",
 ] }
-egui_extras.workspace = true
 egui_plot.workspace = true
 egui-wgpu.workspace = true
 egui.workspace = true
