+++ crates/re_viewer/src/app_state.rs
@@ -11,8 +11,8 @@ use re_viewer_context::{
     ComponentUiRegistry, PlayState, RecordingConfig, SpaceViewClassRegistry, StoreContext,
     StoreHub, SystemCommandSender as _, ViewerContext,
 };
-use re_viewport::{Viewport, ViewportBlueprint, ViewportState};
-use re_viewport_blueprint::{DataQuery as _, PropertyResolver as _};
+use re_viewport::{Viewport, ViewportState};
+use re_viewport_blueprint::{DataQuery as _, PropertyResolver as _, ViewportBlueprint};
 
 use crate::ui::recordings_panel_ui;
 use crate::{app_blueprint::AppBlueprint, ui::blueprint_panel_ui};
