+++ crates/re_viewport/Cargo.toml
@@ -19,7 +19,6 @@ workspace = true
 all-features = true
 
 [dependencies]
-re_data_store.workspace = true
 re_data_ui.workspace = true
 re_entity_db = { workspace = true, features = ["serde"] }
 re_log_types.workspace = true
@@ -35,7 +34,6 @@ re_space_view.workspace = true
 re_tracing.workspace = true
 re_types.workspace = true
 re_types_blueprint.workspace = true
-re_types_core.workspace = true
 re_ui.workspace = true
 re_viewer_context.workspace = true
 re_viewport_blueprint.workspace = true
