+++ crates/re_viewport/src/add_space_view_or_container_modal.rs
@@ -4,9 +4,9 @@ use re_ui::ReUi;
 use re_viewer_context::{
     blueprint_id_to_tile_id, ContainerId, RecommendedSpaceView, ViewerContext,
 };
-use re_viewport_blueprint::SpaceViewBlueprint;
+use re_viewport_blueprint::{SpaceViewBlueprint, ViewportBlueprint};
 
-use crate::{icon_for_container_kind, ViewportBlueprint};
+use crate::icon_for_container_kind;
 
 #[derive(Default)]
 pub struct AddSpaceViewOrContainerModal {
