+++ crates/re_viewport/src/space_view_entity_picker.rs
@@ -6,9 +6,7 @@ use re_entity_db::{EntityPath, EntityTree, InstancePath};
 use re_log_types::{EntityPathFilter, EntityPathRule};
 use re_space_view::determine_visualizable_entities;
 use re_viewer_context::{DataQueryResult, SpaceViewId, ViewerContext};
-use re_viewport_blueprint::SpaceViewBlueprint;
-
-use crate::ViewportBlueprint;
+use re_viewport_blueprint::{SpaceViewBlueprint, ViewportBlueprint};
 
 /// Window for adding/removing entities from a space view.
 ///
