+++ crates/re_viewport_blueprint/Cargo.toml
@@ -25,9 +25,12 @@ re_log_types.workspace = true
 re_log.workspace = true
 re_tracing.workspace = true
 re_types.workspace = true
+re_types_blueprint.workspace = true
 re_types_core.workspace = true
 re_viewer_context.workspace = true
 
+ahash.workspace = true
+egui_tiles.workspace = true
 egui.workspace = true
 itertools.workspace = true
 nohash-hasher.workspace = true
